+++ CHANGELOG.rst
@@ -2,6 +2,16 @@
 Changelog
 ---------
 
+5.0.0
+~~~~~
+
+a) [TASK] Introduce deployer-loader in src/Loader.php
+b) [TASK] Add default value for 'web_path' and set it to empty.
+c) [TASK] Remove 'web_path' from 'clear_paths' for composer.json / composer.lock
+d) [TASK] [!!!BREAKING] Change the way getInstanceName parameters works. Param means now path relative to root of project.
+e) [TASK] [!!!BREAKING] Change the way getDatabaseConfig parameters works. Param means now path relative to root of project.
+f) [TASK] Update dependent packages. Add deployer-loader package.
+
 4.0.1
 ~~~~~